+++ source/src/data-mapper/Model/ConditionType.cs
@@ -5,6 +5,7 @@
         DEFAULT = 0,
         VALUE_EXISTS = 1,
         VALUE_EQUALS = 2,
-        VALUE_DOES_NOT_EQUAL = 3
+        VALUE_DOES_NOT_EQUAL = 3,
+        VALUE_MATCHES = 4
     }
 }